requirements.txt
datasets==3.1.0
numpy==2.2.5
openai==1.75.0
pandas==2.2.3
Pillow==11.2.1
python_Levenshtein==0.27.1
qwen_vl_utils==0.0.11
tqdm==4.67.1
transformers==4.49.0
